I added a new system for abilites, now it is much simpler:
Instead of there being a lot of things that are inherent to classes that
extend Ability (all in the fight.specials package), now you just have ot
have the effects along with the name and MID (magic id)

The MID is so that when a player click,s the listener checks to see if
the lore of the item in hand matches the MID's of any of the abilities
that player's class has. If it does, that specific ability is activated.
THE LORE IS THE PART OF THE ITEM THAT IS CHECKED. I tried to find a way
to attach the <Ability> directly, but ItemMeta only holds Strings in the
form of Display Name and Lore, so I used lore for the operational part.